+++ src/euroeval/dataset_configs/french.py
@@ -154,3 +154,12 @@ MULTILOKO_FR_CONFIG = DatasetConfig(
     languages=[FRENCH],
     unofficial=True,
 )
+
+MULTINRC_FR_CONFIG = DatasetConfig(
+    name="multinrc-fr",
+    pretty_name="MultiNRC-fr",
+    source="EuroEval/multinrc-fr",
+    task=KNOW,
+    languages=[FRENCH],
+    unofficial=True,
+)