+++ src/euroeval/dataset_configs/spanish.py
@@ -170,3 +170,12 @@ MULTILOKO_ES_CONFIG = DatasetConfig(
     languages=[SPANISH],
     unofficial=True,
 )
+
+MULTINRC_ES_CONFIG = DatasetConfig(
+    name="multinrc-es",
+    pretty_name="MultiNRC-es",
+    source="EuroEval/multinrc-es",
+    task=KNOW,
+    languages=[SPANISH],
+    unofficial=True,
+)